% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerce_to_integer}
\alias{coerce_to_integer}
\title{Coerce an object to an integer, or error if that's not possible}
\usage{
coerce_to_integer(x)
}
\arguments{
\item{x}{An integer-like value, or a character that represents an integer.}
}
\value{
an integer if possible, or an error if not
}
\description{
R isn't the best language for strict typing, and it particularly struggles
with integers. For example, \code{is.integer(5)} returns \code{FALSE}, because \code{5} is
not the same as \code{5L}. This function will attempt to convert a numeric or a
character
}
\details{
A simpler (and potentially faster) implementation of this function would use
the \code{\link[rlang]{is_integerish}} function.
}
\keyword{internal}
