% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity.R
\name{is_even}
\alias{is_even}
\alias{is_odd}
\title{Determine if a number is even}
\usage{
is_even(int)

is_odd(int)
}
\arguments{
\item{int}{Integer, or number that can be coerced to integer}
}
\value{
logical
}
\description{
Determine if a number is even
}
