% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity.R
\name{parity}
\alias{parity}
\title{Advanced algorithm for determining if a number is odd or even}
\usage{
parity(int)
}
\arguments{
\item{int}{Integer, or number that can be coerced to integer}
}
\value{
character, one of "even" or "odd"
}
\description{
Advanced algorithm for determining if a number is odd or even
}
\examples{
parity(3) # odd
parity(8) # even
parity(-12) # odd
parity(0) # even

}
